+++ src/interfaces.ts
@@ -106,6 +106,11 @@ export interface JobHelpers extends Helpers {
   abortSignal?: AbortSignal;
 }
 
+export type CleanupTask =
+  | "GC_TASK_IDENTIFIERS"
+  | "GC_JOB_QUEUES"
+  | "DELETE_PERMAFAILED_JOBS";
+
 /**
  * Utilities for working with Graphile Worker. Primarily useful for migrating
  * the jobs database and queueing jobs.
@@ -170,6 +175,19 @@ export interface WorkerUtils extends Helpers {
    * were terminated, are permanently unreachable, etc).
    */
   forceUnlockWorkers: (workerIds: string[]) => Promise<void>;
+
+  /**
+   * **Experimental**
+   *
+   * Database cleanup function. Supported tasks:
+   *
+   * - GC_TASK_IDENTIFIERS: delete task identifiers that are no longer referenced by any jobs
+   * - GC_JOB_QUEUES: delete job queues that are no longer referenced by any jobs
+   * - DELETE_PERMAFAILED_JOBS: delete permanently failed jobs if they are not locked
+   *
+   * Default: ["GC_JOB_QUEUES"]
+   */
+  cleanup(options: { tasks?: CleanupTask[] }): Promise<void>;
 }
 
 export type PromiseOrDirect<T> = Promise<T> | T;